+++ src/utils/excel.ts
@@ -98,6 +98,7 @@ export const bouwBackupWerkboek = (data: any): XLSX.WorkBook => {
     )
     addSheet('TankReinigingLog',      data.tank_reinigingslog)
     addSheet('Artikelen',             data.artikelen)
+    addSheet('DropshipArtikelen',     data.dropship_artikelen)
     addSheet('HygieneItems',          data.hygiene_items)
     addSheet('HygieneGroups',         data.hygiene_groups)
     addSheet('BrouwdagChecklist',     data.brouwdag_checklist)
@@ -294,6 +295,7 @@ export const parseBackupWerkboek = (wb: XLSX.WorkBook): any => {
         })(),
         tank_reinigingslog:           parse('TankReinigingLog'),
         artikelen:                    parse('Artikelen'),
+        dropship_artikelen:           parse('DropshipArtikelen'),
         hygiene_items:                parse('HygieneItems'),
         hygiene_groups:               parse('HygieneGroups'),
         brouwdag_checklist:           parse('BrouwdagChecklist'),